+++ CHANGELOG.md
@@ -78,3 +78,4 @@
 * (ERICK) Updated student class structure to match staff structure.
 * (ERICK) Updated module class structure to match staff structure.
 * (ERICK) `reserveOptions()` now loops until choosing to exit.
+* (ERICK) Module booking is correctly mapped to module.